Day one
Go-Kart platformer game with the objective to make it to the Sanctum of Speed. Climb mountains with the risk of falling off at any time and starting from the beginning of the mountain. After reaching the top there will be a bridge to take you to the next mountain's island. Progressing further will allow you to improve your kart.